Feat : 인게임 피어 호스트 기능을 구현

+++ Assets/Scripts/Lobby/LobbyReadyButton.cs
@@ -57,7 +57,7 @@ public class LobbyReadyButton : MonoBehaviour
     {
         if (NetManager.Instance == null)
             return;
-        if ((int)packet.PlayerId != NetManager.Instance._playerId)
+        if (packet.PlayerId != NetManager.Instance._playerId)
             return;
 
         _localReady = packet.IsReady;
